feat: Separate the web server logic in shared_agents_utils.py into a separate module. This will need to be imported in other files
The web server logic (ApprovalWebServer, ApprovalHandler, and wait_for_user_approval_from_browser) is currently embedded in shared_agents_utils.py. To improve modularity and separation of concerns, this logic will be extracted into a new dedicated module, web_server_utils.py. Both shared_agents_utils.py and code_agent.py currently use these components, so they will be updated to import from the new module. The generation order ensures the new module is created before other files are modified to import from it.
